GameOver functionality
Added obstacle image to asset folder. Added spawn logic and trigger for 3 obstacles to spawn at the start of the game. To let the player know the game has ended and that they died I made it so the player turns red on contact, input gets disabled and "Game Over" text will be displayed. The game will wait 3 seconds to reload a new game automatically as of right now.